Add PlantList component, update App
